Move common code into CryptographyManager.
